Rename Myprofile file
